changes 25/4/24

+++ src/main/java/com/avitam/fantasy11/model/PointsMasterRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PointsMasterRepository extends MongoRepository<PointsMaster,Integer> {
+public interface PointsMasterRepository extends MongoRepository<PointsMaster, Long> {
 }
