+++ src/main/java/com/avitam/fantasy11/model/UserWinningsRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserWinningsRepository extends MongoRepository<UserWinnings,Integer> {
+public interface UserWinningsRepository extends MongoRepository<UserWinnings, Long> {
 }
